pyproject.toml
[tool.poetry]
name = "deepwater3"
version = "0.1.0"
description = "Robust image steganography with test-driven reliability"
authors = ["DJStompZone <stompzone@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Pillow = "^10.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
